Python/Dashboard/requirements.txt
Flask==0.10.1
Flask_SQLAlchemy==1.0
coloredlogs==7.3
flask_bcrypt==0.7.1
flask_login==0.4.1
paho_mqtt==1.3.1
WTForms==2.1
